requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
strands-agents[ollama]==0.3.0
strands-agents-tools==0.1.3
pydantic==2.10.3
httpx==0.27.2